commit2

+++ app/Http/Controllers/TransaksiController.php
@@ -2,10 +2,12 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\DetailBelanja;
 use App\Models\HeaderBelanja;
 use App\Models\Item;
 use App\Models\Karyawan;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Redirect;
 use Inertia\Inertia;
 
 class TransaksiController extends Controller
@@ -29,17 +31,53 @@ class TransaksiController extends Controller
     public function store(Request $request)
     {
         // dd($request->all());
+        date_default_timezone_set('Asia/Jakarta');
         $data = [
-            'tanggal_belanja' => $request->tanggal_belanja,
+            'tanggal_belanja' => date('Y-m-d H:i:s'),
             'id_karyawan' => $request->id_karyawan,
-            'total_belanja' => $request->total_belanja
+            'total_belanja' => $request->total_belanja,
         ];
+
+        // $dataDetail = [
+        //     'id_belanja' => $request->id,
+        //     'id_item' => $request->id_item,
+        //     'quantity' => $request->quantity,
+        //     'jumlah' => $request->jumlah
+        // ];
 
         $transaction = HeaderBelanja::create($data);
 
         if ($transaction) {
-            return redirect('transaction')->with('message', 'Transaksi Berhasil Disimpan!');
+            $details = $request->detail;
+
+            if (count($details) > 0) {
+                foreach ($details as $detail) {
+                    $dataDetail = [
+                        'id_belanja' => $transaction->id,
+                        'id_item' => $detail['item'],
+                        'quantity' => $detail['quantity'],
+                        'jumlah' => $detail['jumlah'],
+                    ];
+
+                    DetailBelanja::create($dataDetail);
+                }
+            }
+            return Redirect::route('transactions.index')->with('message', 'Data Berhasil Disimpan!');
+
+            // return redirect('transaction')->with('message', 'Transaksi Berhasil Disimpan!');
         }
+    }
+
+    public function show(HeaderBelanja $transaction)
+    {
+        $jumlahTotal = 0;
+        foreach ($transaction->items as $item) {
+            $jumlahTotal += $item->pivot->jumlah;
+        }
+        return Inertia::render('Transaksi/Detail', [
+            'details' => $transaction->items,
+            'jumlahTotal' => $jumlahTotal,
+        ]);
     }
 
     public function edit(Request $request, HeaderBelanja $transaction)
@@ -63,7 +101,12 @@ class TransaksiController extends Controller
 
     public function destroy(Request $request, HeaderBelanja $transaction)
     {
-        $transaction->delete();
+
+        $deleteTx = DetailBelanja::where('id_belanja', $transaction->id)->delete();
+        if ($deleteTx) {
+            $transaction->delete();
+        }
+
         return redirect('transactions')->with('message', 'Transaksi Berhasil Di Hapus!');
     }
 }
